refactor(sdk): move `load_linked_chunk_metadata` to a file where it can be reused for other linked chunks
Makes sense to me to have it in the persistence file, since it's
reloading all the chunks quickly to set the initial ordering.
